Moved dashboard to views folder
